Audio: clean sync start event when initilizing RecordThread
Issue Description:
Remote side is muted after killed mediaserver.

Root Cause:
When killed mediaserver, it will recreate RecordThread.
It will check the mFramestoDrop to use SyncEvent. But
in construction of RecordThread, mFramestoDrop is not
initilized as 0. So this value could be anything. And
RecordThread will be blocked as mFramestoDrop.

Solution:
Clean sync start event when initilizing RecordThread.
And let user to choose whether to use SyncEvent in
start() function.

Change-Id: Id279a638f40ad965fc81060e7929acee4354fcbd
Signed-off-by: Chao Deng <cdeng@marvell.com>